Add files via upload
pomolife4. 
I added a plus button to it that regroups the pre-existing input field for a new task and a new value dropdown list. The user can toggle the plus button. 

The value selected by user will be added to the end of the new task via a <p> tag.